second lab

+++ src/Point.java
@@ -1,8 +1,8 @@
 public class Point {
-    public final int x;
-    public final int y;
+    public final double x;
+    public final double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
